% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascarade-package.R
\docType{data}
\name{exampleMascarade}
\alias{exampleMascarade}
\title{Example data with UMAP points from PBMC3K dataset.}
\description{
The object is a list with three elements:
1) `dims` -- matrix of UMAP coordinates of the cells,
2) `clusters` -- vector of cell population annotations,
3) `features` -- matrix withgene expression for several genes.
}
