% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mascarade-package.R
\docType{data}
\name{exampleMaskTable}
\alias{exampleMaskTable}
\title{Example generated mask.}
\description{
Result of \code{generateMasks(dims=exampleMascarade$dims, clusters=exampleMascarade$clusters)}.
}
